# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Office management control system API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Log in with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
...
  /clients:
    get:
      operationId: listClients
      tags: [clients]
      parameters:
        - in: query
          name: search
          schema: { type: string }
        - in: query
          name: status
          schema:
            type: string
            enum: [active, inactive, prospect]
      responses:
        "200":
          description: Clients
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Client" }
    post:
      operationId: createClient
      tags: [clients]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateClientBody" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Client" }
  /clients/{id}:
    get:
      operationId: getClient
      tags: [clients]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Client
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Client" }
    patch:
      operationId: updateClient
      tags: [clients]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateClientBody" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Client" }
    delete:
      operationId: deleteClient
      tags: [clients]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ---------- Tasks ----------
...
  /attendance:
    get:
      operationId: listAttendance
      tags: [attendance]
      parameters:
        - in: query
          name: staffId
          schema: { type: integer }
        - in: query
          name: from
          schema: { type: string, format: date }
        - in: query
          name: to
          schema: { type: string, format: date }
      responses:
        "200":
          description: Attendance entries
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/AttendanceEntry" }
  /attendance/check-in:
    post:
      operationId: checkIn
      tags: [attendance]
      summary: Check in current user
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CheckInBody" }
      responses:
        "200":
          description: Checked in
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AttendanceEntry" }
  /attendance/check-out:
    post:
      operationId: checkOut
      tags: [attendance]
      summary: Check out current user
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CheckOutBody" }
      responses:
        "200":
          description: Checked out
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AttendanceEntry" }
...
  /invoices:
    get:
      operationId: listInvoices
      tags: [invoices]
      parameters:
        - in: query
          name: clientId
          schema: { type: integer }
        - in: query
          name: status
          schema:
            type: string
            enum: [draft, sent, paid, overdue, cancelled]
      responses:
        "200":
          description: Invoices
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Invoice" }
    post:
      operationId: createInvoice
      tags: [invoices]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateInvoiceBody" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Invoice" }
  /invoices/{id}:
    get:
      operationId: getInvoice
      tags: [invoices]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Invoice
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Invoice" }
    patch:
      operationId: updateInvoice
      tags: [invoices]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateInvoiceBody" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Invoice" }
    delete:
      operationId: deleteInvoice
      tags: [invoices]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ---------- Notifications ----------
...
components:
  schemas:
...
    LoginBody:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }
    AuthUser:
      type: object
      required: [id, email, name, role]
      properties:
        id: { type: integer }
        email: { type: string }
        name: { type: string }
        role:
          type: string
          enum: [admin, staff]
        position: { type: string, nullable: true }
        phone: { type: string, nullable: true }
        department: { type: string, nullable: true }
        avatarUrl: { type: string, nullable: true }
...
    Client:
      type: object
      required: [id, name, status, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        company: { type: string, nullable: true }
        email: { type: string, nullable: true }
        phone: { type: string, nullable: true }
        address: { type: string, nullable: true }
        status:
          type: string
          enum: [active, inactive, prospect]
        notes: { type: string, nullable: true }
        contactPerson: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
    CreateClientBody:
      type: object
      required: [name, status]
      properties:
        name: { type: string, minLength: 1 }
        company: { type: string, nullable: true }
        email: { type: string, nullable: true }
        phone: { type: string, nullable: true }
        address: { type: string, nullable: true }
        status:
          type: string
          enum: [active, inactive, prospect]
        notes: { type: string, nullable: true }
        contactPerson: { type: string, nullable: true }
    UpdateClientBody:
      type: object
      properties:
        name: { type: string }
        company: { type: string, nullable: true }
        email: { type: string, nullable: true }
        phone: { type: string, nullable: true }
        address: { type: string, nullable: true }
        status:
          type: string
          enum: [active, inactive, prospect]
        notes: { type: string, nullable: true }
        contactPerson: { type: string, nullable: true }
...
    AttendanceEntry:
      type: object
      required: [id, staffId, staffName, date, status]
      properties:
        id: { type: integer }
        staffId: { type: integer }
        staffName: { type: string }
        date: { type: string, format: date }
        checkIn: { type: string, nullable: true, format: date-time }
        checkOut: { type: string, nullable: true, format: date-time }
        status:
          type: string
          enum: [present, absent, late, on_leave]
        notes: { type: string, nullable: true }
        hoursWorked: { type: number, nullable: true }
    CheckInBody:
      type: object
      properties:
        notes: { type: string, nullable: true }
    CheckOutBody:
      type: object
      properties:
        notes: { type: string, nullable: true }
...
    InvoiceItem:
      type: object
      required: [description, quantity, unitPrice]
      properties:
        description: { type: string }
        quantity: { type: number }
        unitPrice: { type: number }
    Invoice:
      type: object
      required: [id, invoiceNumber, clientId, clientName, status, total, issueDate, createdAt]
      properties:
        id: { type: integer }
        invoiceNumber: { type: string }
        clientId: { type: integer }
        clientName: { type: string }
        status:
          type: string
          enum: [draft, sent, paid, overdue, cancelled]
        items:
          type: array
          items: { $ref: "#/components/schemas/InvoiceItem" }
        subtotal: { type: number }
        tax: { type: number }
        total: { type: number }
        issueDate: { type: string, format: date }
        dueDate: { type: string, nullable: true, format: date }
        paidAt: { type: string, nullable: true, format: date-time }
        notes: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
    CreateInvoiceBody:
      type: object
      required: [clientId, items, issueDate, status]
      properties:
        clientId: { type: integer }
        status:
          type: string
          enum: [draft, sent, paid, overdue, cancelled]
        items:
          type: array
          items: { $ref: "#/components/schemas/InvoiceItem" }
        tax: { type: number, nullable: true }
        issueDate: { type: string, format: date }
        dueDate: { type: string, nullable: true, format: date }
        notes: { type: string, nullable: true }
    UpdateInvoiceBody:
      type: object
      properties:
        clientId: { type: integer }
        status:
          type: string
          enum: [draft, sent, paid, overdue, cancelled]
        items:
          type: array
          items: { $ref: "#/components/schemas/InvoiceItem" }
        tax: { type: number, nullable: true }
        issueDate: { type: string, format: date }
        dueDate: { type: string, nullable: true, format: date }
        notes: { type: string, nullable: true }
...