# lib/api-zod/src/generated/types/attendanceEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Office management control system API
 * OpenAPI spec version: 0.1.0
 */
import type { AttendanceEntryStatus } from "./attendanceEntryStatus";

export interface AttendanceEntry {
  id: number;
  staffId: number;
  staffName: string;
  date: Date;
  checkIn?: Date | null;
  checkOut?: Date | null;
  status: AttendanceEntryStatus;
  notes?: string | null;
  hoursWorked?: number | null;
}