# lib/api-zod/src/generated/types/createInvoiceBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Office management control system API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateInvoiceBodyStatus } from "./createInvoiceBodyStatus";
import type { InvoiceItem } from "./invoiceItem";

export interface CreateInvoiceBody {
  clientId: number;
  status: CreateInvoiceBodyStatus;
  items: InvoiceItem[];
  tax?: number | null;
  issueDate: Date;
  dueDate?: Date | null;
  notes?: string | null;
}